+++ lib/src/button/sbb_primary_button.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 
 import '../../design_system_flutter.dart';
-import '../sbb_internal.dart';
 
 /// The SBB Primary Button. Use according to documentation.
 ///
@@ -18,15 +17,13 @@ import '../sbb_internal.dart';
 class SBBPrimaryButton extends StatelessWidget {
   const SBBPrimaryButton({
     super.key,
-    this.label,
-    this.child,
+    required this.label,
     this.isLoading = false,
     required this.onPressed,
     this.focusNode,
-  }) : assert(label != null && child == null || label == null && child != null);
+  });
 
-  final String? label;
-  final Widget? child;
+  final String label;
   final bool isLoading;
   final VoidCallback? onPressed;
   final FocusNode? focusNode;
@@ -44,29 +41,31 @@ class SBBPrimaryButton extends StatelessWidget {
   }
 
   ElevatedButton _buildThemedMobile(BuildContext context) {
+    final styles = SBBButtonStyles.of(context);
     return ElevatedButton(
-      style: SBBButtonStyles.of(context).primaryMobile,
+      style: styles.primaryMobile,
       onPressed: isLoading ? null : onPressed,
       focusNode: focusNode,
       child: Row(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
           if (isLoading) SBBLoadingIndicator.tinyCloud(),
-          if (label == null) child! else SBBButtonContent(label: label!),
+          styles.buttonLabelBuilder!(context, label),
         ],
       ),
     );
   }
 
   ElevatedButton _buildThemedWeb(BuildContext context) {
+    final styles = SBBButtonStyles.of(context);
     return ElevatedButton(
-      style: SBBButtonStyles.of(context).primaryWebLean,
+      style: styles.primaryWebLean,
       onPressed: onPressed,
       focusNode: focusNode,
       child: Row(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
-          if (label == null) child! else SBBButtonContent(label: label!),
+          styles.buttonLabelBuilder!(context, label),
         ],
       ),
     );
@@ -101,53 +100,42 @@ class SBBPrimaryButtonNegative extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final theme = Theme.of(context);
     switch (SBBBaseStyle.of(context).hostPlatform) {
       case HostPlatform.native:
-        return _buildThemedMobile(theme);
+        return _buildThemedMobile(context);
       case HostPlatform.web:
-        return _buildThemedWeb(theme);
+        return _buildThemedWeb(context);
       default:
-        return _buildThemedMobile(theme);
+        return _buildThemedMobile(context);
     }
   }
 
-  Widget _buildThemedMobile(ThemeData theme) {
+  Widget _buildThemedMobile(BuildContext context) {
+    final styles = SBBButtonStyles.of(context);
     return ElevatedButton(
-      style: theme.extension<SBBButtonStyles>()?.primaryMobileNegative,
+      style: styles.primaryMobileNegative,
       onPressed: isLoading ? null : onPressed,
       focusNode: focusNode,
       child: Row(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
           if (isLoading) SBBLoadingIndicator.tinyCloud(),
-          Flexible(
-            child: Text(
-              label,
-              overflow: TextOverflow.ellipsis,
-              maxLines: 1,
-            ),
-          ),
+          styles.buttonLabelBuilder!(context, label),
         ],
       ),
     );
   }
 
-  Widget _buildThemedWeb(ThemeData theme) {
+  Widget _buildThemedWeb(BuildContext context) {
+    final styles = SBBButtonStyles.of(context);
     return ElevatedButton(
-      style: theme.extension<SBBButtonStyles>()?.primaryWebNegative,
+      style: styles.primaryWebNegative,
       onPressed: onPressed,
       focusNode: focusNode,
       child: Row(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
-          Flexible(
-            child: Text(
-              label,
-              overflow: TextOverflow.ellipsis,
-              maxLines: 1,
-            ),
-          ),
+          styles.buttonLabelBuilder!(context, label),
         ],
       ),
     );