+++ lib/src/status/sbb_status.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 
 const _kBorderRadius = 2.0;
 
+@Deprecated('will soon be moved to the web package')
 class SBBStatus extends StatelessWidget {
   const SBBStatus.valid({
     Key? key,